+++ sdk/communication/azure-communication-jobrouter/src/main/java/com/azure/communication/jobrouter/JobRouterAdministrationClient.java
@@ -4,10 +4,6 @@
 package com.azure.communication.jobrouter;
 
 import com.azure.communication.jobrouter.implementation.JobRouterAdministrationClientImpl;
-import com.azure.communication.jobrouter.implementation.accesshelpers.ClassificationPolicyConstructorProxy;
-import com.azure.communication.jobrouter.implementation.accesshelpers.DistributionPolicyConstructorProxy;
-import com.azure.communication.jobrouter.implementation.accesshelpers.ExceptionPolicyConstructorProxy;
-import com.azure.communication.jobrouter.implementation.accesshelpers.RouterQueueConstructorProxy;
 import com.azure.communication.jobrouter.implementation.converters.ClassificationPolicyAdapter;
 import com.azure.communication.jobrouter.implementation.converters.DistributionPolicyAdapter;
 import com.azure.communication.jobrouter.implementation.converters.ExceptionPolicyAdapter;
@@ -303,9 +299,8 @@ public final class JobRouterAdministrationClient {
         Response<BinaryData> response = this.serviceClient.upsertDistributionPolicyWithResponse(
             createDistributionPolicyOptions.getDistributionPolicyId(), BinaryData.fromObject(distributionPolicy),
             requestOptions);
-        return new SimpleResponse<DistributionPolicy>(response.getRequest(), response.getStatusCode(),
-            response.getHeaders(),
-            DistributionPolicyConstructorProxy.create(response.getValue().toObject(DistributionPolicyInternal.class)));
+        return new SimpleResponse<>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
+            response.getValue().toObject(DistributionPolicy.class));
     }
 
     /**
@@ -705,9 +700,8 @@ public final class JobRouterAdministrationClient {
         Response<BinaryData> response = this.serviceClient.upsertClassificationPolicyWithResponse(
             createClassificationPolicyOptions.getClassificationPolicyId(), BinaryData.fromObject(classificationPolicy),
             requestOptions);
-        return new SimpleResponse<ClassificationPolicy>(response.getRequest(), response.getStatusCode(),
-            response.getHeaders(), ClassificationPolicyConstructorProxy
-                .create(response.getValue().toObject(ClassificationPolicyInternal.class)));
+        return new SimpleResponse<>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
+            response.getValue().toObject(ClassificationPolicy.class));
     }
 
     /**
@@ -1114,9 +1108,8 @@ public final class JobRouterAdministrationClient {
         Response<BinaryData> response
             = this.serviceClient.upsertExceptionPolicyWithResponse(createExceptionPolicyOptions.getExceptionPolicyId(),
                 BinaryData.fromObject(exceptionPolicy), requestOptions);
-        return new SimpleResponse<ExceptionPolicy>(response.getRequest(), response.getStatusCode(),
-            response.getHeaders(),
-            ExceptionPolicyConstructorProxy.create(response.getValue().toObject(ExceptionPolicyInternal.class)));
+        return new SimpleResponse<>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
+            response.getValue().toObject(ExceptionPolicy.class));
     }
 
     /**
@@ -1482,8 +1475,8 @@ public final class JobRouterAdministrationClient {
         RouterQueueInternal queue = QueueAdapter.convertCreateQueueOptionsToRouterQueueInternal(createQueueOptions);
         Response<BinaryData> response = this.serviceClient.upsertQueueWithResponse(createQueueOptions.getQueueId(),
             BinaryData.fromObject(queue), requestOptions);
-        return new SimpleResponse<RouterQueue>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
-            RouterQueueConstructorProxy.create(response.getValue().toObject(RouterQueueInternal.class)));
+        return new SimpleResponse<>(response.getRequest(), response.getStatusCode(), response.getHeaders(),
+            response.getValue().toObject(RouterQueue.class));
     }
 
     /**
@@ -1614,7 +1607,6 @@ public final class JobRouterAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public DistributionPolicy getDistributionPolicy(String distributionPolicyId) {
-        // Generated convenience method for getDistributionPolicyWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getDistributionPolicyWithResponse(distributionPolicyId, requestOptions).getValue()
             .toObject(DistributionPolicy.class);
@@ -1671,7 +1663,6 @@ public final class JobRouterAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public ClassificationPolicy getClassificationPolicy(String classificationPolicyId) {
-        // Generated convenience method for getClassificationPolicyWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getClassificationPolicyWithResponse(classificationPolicyId, requestOptions).getValue()
             .toObject(ClassificationPolicy.class);
@@ -1728,7 +1719,6 @@ public final class JobRouterAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public ExceptionPolicy getExceptionPolicy(String exceptionPolicyId) {
-        // Generated convenience method for getExceptionPolicyWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getExceptionPolicyWithResponse(exceptionPolicyId, requestOptions).getValue()
             .toObject(ExceptionPolicy.class);
@@ -1746,7 +1736,6 @@ public final class JobRouterAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
     public PagedIterable<ExceptionPolicy> listExceptionPolicies() {
-        // Generated convenience method for listExceptionPolicies
         RequestOptions requestOptions = new RequestOptions();
         return serviceClient.listExceptionPolicies(requestOptions)
             .mapPage(bodyItemValue -> bodyItemValue.toObject(ExceptionPolicy.class));
@@ -1785,7 +1774,6 @@ public final class JobRouterAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public RouterQueue getQueue(String queueId) {
-        // Generated convenience method for getQueueWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getQueueWithResponse(queueId, requestOptions).getValue().toObject(RouterQueue.class);
     }
@@ -1802,7 +1790,6 @@ public final class JobRouterAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
     public PagedIterable<RouterQueue> listQueues() {
-        // Generated convenience method for listQueues
         RequestOptions requestOptions = new RequestOptions();
         return serviceClient.listQueues(requestOptions)
             .mapPage(bodyItemValue -> bodyItemValue.toObject(RouterQueue.class));