+++ sdk/communication/azure-communication-jobrouter/src/main/java/com/azure/communication/jobrouter/models/StaticRouterRule.java
@@ -7,6 +7,8 @@ import com.azure.core.annotation.Fluent;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 /** A rule providing static rules that always return the same result, regardless of input. */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "kind")
@@ -18,6 +20,8 @@ public final class StaticRouterRule extends RouterRule {
      * The static value this rule always returns.
      */
     @JsonProperty(value = "value")
+    @JsonDeserialize(using = RouterValueDeserializer.class)
+    @JsonSerialize(using = RouterValueSerializer.class)
     private RouterValue value;
 
     /** Creates an instance of StaticRouterRule class. */