+++ sdk/communication/azure-communication-jobrouter/src/test/java/com.azure.communication.jobrouter/ClassificationPolicyAsyncLiveTests.java
@@ -22,7 +22,9 @@ import com.azure.communication.jobrouter.models.StaticRouterRule;
 import com.azure.communication.jobrouter.models.StaticWorkerSelectorAttachment;
 import com.azure.communication.jobrouter.models.WorkerSelectorAttachment;
 import com.azure.core.http.HttpClient;
+import com.azure.core.http.rest.Response;
 import com.azure.core.test.TestMode;
+import com.azure.core.util.BinaryData;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -33,6 +35,8 @@ import java.util.List;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class ClassificationPolicyAsyncLiveTests extends JobRouterTestBase {
     private JobRouterAdministrationAsyncClient administrationAsyncClient;
@@ -90,7 +94,8 @@ public class ClassificationPolicyAsyncLiveTests extends JobRouterTestBase {
         };
 
         StaticWorkerSelectorAttachment staticWorkerSelector = new StaticWorkerSelectorAttachment(
-            new RouterWorkerSelector("key", LabelOperator.EQUAL, new RouterValue("value")));
+            new RouterWorkerSelector("key", LabelOperator.EQUAL, new RouterValue("value"))
+                .setExpedite(true).setExpiresAfter(Duration.ofSeconds(10)));
 
         List<WorkerSelectorAttachment> workerSelectors = new ArrayList<WorkerSelectorAttachment>() {
             {
@@ -110,7 +115,7 @@ public class ClassificationPolicyAsyncLiveTests extends JobRouterTestBase {
         String channelId = String.format("%s-%s-Channel", JAVA_LIVE_TESTS, testName);
 
         // Action
-        ClassificationPolicy result = administrationAsyncClient.createClassificationPolicy(createClassificationPolicyOptions).block();
+        ClassificationPolicy policy = administrationAsyncClient.createClassificationPolicy(createClassificationPolicyOptions).block();
         RouterJob job = routerAsyncClient.createJobWithClassificationPolicy(
             new CreateJobWithClassificationPolicyOptions(jobId, channelId, classificationPolicyId)).block();
 
@@ -119,12 +124,30 @@ public class ClassificationPolicyAsyncLiveTests extends JobRouterTestBase {
         }
 
         // Verify
-        assertEquals(classificationPolicyId, result.getId());
-        assertEquals(classificationPolicyName, result.getName());
-        assertEquals(StaticRouterRule.class, result.getPrioritizationRule().getClass());
-        assertEquals(1, result.getWorkerSelectorAttachments().size());
-        assertEquals(1, result.getQueueSelectorAttachments().size());
-        assertEquals(fallbackQueueId, result.getFallbackQueueId());
+        assertEquals(classificationPolicyId, policy.getId());
+        assertEquals(classificationPolicyName, policy.getName());
+        assertEquals(StaticRouterRule.class, policy.getPrioritizationRule().getClass());
+        assertNotNull(policy.getEtag());
+        assertEquals(1, ((StaticRouterRule) policy.getPrioritizationRule()).getValue().getIntValue());
+        assertEquals(1, policy.getWorkerSelectorAttachments().size());
+        assertEquals(Duration.ofSeconds(10), ((StaticWorkerSelectorAttachment) policy.getWorkerSelectorAttachments().get(0)).getWorkerSelector().getExpiresAfter());
+        assertEquals(1, policy.getQueueSelectorAttachments().size());
+        assertEquals(fallbackQueueId, policy.getFallbackQueueId());
+
+        Response<BinaryData> binaryData = administrationAsyncClient.getClassificationPolicyWithResponse(policy.getId(), null).block();
+        ClassificationPolicy deserialized = binaryData.getValue().toObject(ClassificationPolicy.class);
+
+        assertEquals(classificationPolicyId, deserialized.getId());
+        assertEquals(classificationPolicyName, deserialized.getName());
+        assertEquals(StaticRouterRule.class, deserialized.getPrioritizationRule().getClass());
+        assertEquals(policy.getEtag(), deserialized.getEtag());
+        assertEquals(1, ((StaticRouterRule) deserialized.getPrioritizationRule()).getValue().getIntValue());
+        assertEquals(1, deserialized.getWorkerSelectorAttachments().size());
+        assertEquals(Duration.ofSeconds(10), ((StaticWorkerSelectorAttachment) deserialized.getWorkerSelectorAttachments().get(0)).getWorkerSelector().getExpiresAfter());
+        assertEquals(1, deserialized.getQueueSelectorAttachments().size());
+        assertEquals(fallbackQueueId, deserialized.getFallbackQueueId());
+
+        job = routerAsyncClient.getJob(job.getId()).block();
 
         assertEquals(jobId, job.getId());
         assertEquals(classificationPolicyId, job.getClassificationPolicyId());
@@ -132,6 +155,21 @@ public class ClassificationPolicyAsyncLiveTests extends JobRouterTestBase {
         assertEquals(channelId, job.getChannelId());
         assertEquals(1, job.getPriority());
         assertEquals(1, job.getAttachedWorkerSelectors().size());
+
+        deserialized.setPrioritizationRule(null);
+        deserialized.setQueueSelectorAttachments(new ArrayList<>());
+        ClassificationPolicy updatedPolicy = administrationAsyncClient.updateClassificationPolicy(
+            deserialized.getId(), BinaryData.fromObject(deserialized), null).block().toObject(ClassificationPolicy.class);
+
+        assertEquals(classificationPolicyId, updatedPolicy.getId());
+        assertEquals(classificationPolicyName, updatedPolicy.getName());
+        assertEquals(StaticRouterRule.class, updatedPolicy.getPrioritizationRule().getClass());
+        assertNotEquals(policy.getEtag(), updatedPolicy.getEtag());
+        assertEquals(1, ((StaticRouterRule) updatedPolicy.getPrioritizationRule()).getValue().getIntValue());
+        assertEquals(1, updatedPolicy.getWorkerSelectorAttachments().size());
+        assertEquals(Duration.ofSeconds(10), ((StaticWorkerSelectorAttachment) updatedPolicy.getWorkerSelectorAttachments().get(0)).getWorkerSelector().getExpiresAfter());
+        assertEquals(0, updatedPolicy.getQueueSelectorAttachments().size());
+        assertEquals(fallbackQueueId, updatedPolicy.getFallbackQueueId());
 
         // Cleanup
         routerAsyncClient.cancelJob(job.getId()).block();