+++ sdk/communication/azure-communication-jobrouter/src/test/java/com.azure.communication.jobrouter/RouterJobLiveTests.java
@@ -16,6 +16,7 @@ import com.azure.communication.jobrouter.models.RouterJobNote;
 import com.azure.communication.jobrouter.models.RouterJobOffer;
 import com.azure.communication.jobrouter.models.RouterJobStatus;
 import com.azure.communication.jobrouter.models.RouterQueue;
+import com.azure.communication.jobrouter.models.RouterQueueStatistics;
 import com.azure.communication.jobrouter.models.RouterValue;
 import com.azure.communication.jobrouter.models.RouterWorker;
 import com.azure.communication.jobrouter.models.RouterWorkerSelector;
@@ -23,10 +24,13 @@ import com.azure.communication.jobrouter.models.ScheduleAndSuspendMode;
 import com.azure.communication.jobrouter.models.UnassignJobResult;
 import com.azure.core.http.HttpClient;
 import com.azure.core.http.rest.RequestOptions;
+import com.azure.core.http.rest.Response;
+import com.azure.core.test.TestMode;
 import com.azure.core.util.BinaryData;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
@@ -35,6 +39,7 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class RouterJobLiveTests extends JobRouterTestBase {
     private JobRouterClient jobRouterClient;
@@ -43,7 +48,7 @@ public class RouterJobLiveTests extends JobRouterTestBase {
 
     @ParameterizedTest
     @MethodSource("com.azure.core.test.TestBase#getHttpClients")
-    public void unassignJob(HttpClient httpClient) {
+    public void unassignJob(HttpClient httpClient) throws InterruptedException {
         // Setup
         jobRouterClient = getRouterClient(httpClient);
         routerAdminClient = getRouterAdministrationClient(httpClient);
@@ -91,6 +96,7 @@ public class RouterJobLiveTests extends JobRouterTestBase {
                     put("IntLabel", new RouterValue(10));
                     put("BoolLabel", new RouterValue(true));
                     put("StringLabel", new RouterValue("test"));
+                    put("DoubleValue", new RouterValue(5.3));
                 }
             })
             .setTags(new HashMap<String, RouterValue>() {
@@ -102,7 +108,8 @@ public class RouterJobLiveTests extends JobRouterTestBase {
             })
             .setRequestedWorkerSelectors(new ArrayList<RouterWorkerSelector>() {
                 {
-                    add(new RouterWorkerSelector("IntKey", LabelOperator.GREATER_THAN, new RouterValue(2)));
+                    add(new RouterWorkerSelector("IntKey", LabelOperator.GREATER_THAN, new RouterValue(2))
+                        .setExpedite(true).setExpiresAfter(Duration.ofSeconds(100)));
                     add(new RouterWorkerSelector("BoolKey", LabelOperator.EQUAL, new RouterValue(true)));
                 }
             })
@@ -120,13 +127,61 @@ public class RouterJobLiveTests extends JobRouterTestBase {
         // Verify
         assertEquals(jobId, job.getId());
         assertNotNull(job.getEtag());
-        assertEquals(3, job.getLabels().size());
+        assertEquals(4, job.getLabels().size());
         assertEquals(3, job.getTags().size());
         assertEquals(1, job.getNotes().size());
         assertEquals("code1", job.getDispositionCode());
         assertEquals("ref", job.getChannelReference());
         assertEquals(5, job.getPriority());
         assertEquals(2, job.getRequestedWorkerSelectors().size());
+        assertEquals(Duration.ofSeconds(100), job.getRequestedWorkerSelectors().get(0).getExpiresAfter());
+
+        Response<BinaryData> jobResponse = jobRouterClient.getJobWithResponse(job.getId(), null);
+        RouterJob jobDeserialized = jobResponse.getValue().toObject(RouterJob.class);
+
+        // Verify
+        assertEquals(jobId, jobDeserialized.getId());
+        assertNotNull(jobDeserialized.getEtag());
+        assertEquals(4, jobDeserialized.getLabels().size());
+        assertEquals(3, jobDeserialized.getTags().size());
+        assertEquals(1, jobDeserialized.getNotes().size());
+        assertEquals("code1", jobDeserialized.getDispositionCode());
+        assertEquals("ref", jobDeserialized.getChannelReference());
+        assertEquals(5, jobDeserialized.getPriority());
+        assertEquals(2, jobDeserialized.getRequestedWorkerSelectors().size());
+        assertEquals(Duration.ofSeconds(100), jobDeserialized.getRequestedWorkerSelectors().get(0).getExpiresAfter());
+
+        if (this.getTestMode() != TestMode.PLAYBACK) {
+            Thread.sleep(2000);
+        }
+
+        jobDeserialized.setPriority(10);
+        RouterJob updatedJob = jobRouterClient.updateJob(jobId, BinaryData.fromObject(jobDeserialized), null).toObject(RouterJob.class);
+
+        // Verify
+        assertEquals(jobId, updatedJob.getId());
+        assertNotNull(updatedJob.getEtag());
+        assertEquals(4, updatedJob.getLabels().size());
+        assertEquals(3, updatedJob.getTags().size());
+        assertEquals(1, updatedJob.getNotes().size());
+        assertEquals("code1", updatedJob.getDispositionCode());
+        assertEquals("ref", updatedJob.getChannelReference());
+        assertEquals(10, updatedJob.getPriority());
+        assertEquals(2, updatedJob.getRequestedWorkerSelectors().size());
+        assertEquals(Duration.ofSeconds(100), updatedJob.getRequestedWorkerSelectors().get(0).getExpiresAfter());
+
+        for (RouterJob listJob : jobRouterClient.listJobs(null, queueId, channelId, null, null, null)) {
+            assertEquals(jobId, listJob.getId());
+            assertNotNull(listJob.getEtag());
+            assertEquals(4, listJob.getLabels().size());
+            assertEquals(3, listJob.getTags().size());
+            assertEquals(1, listJob.getNotes().size());
+            assertEquals("code1", listJob.getDispositionCode());
+            assertEquals("ref", listJob.getChannelReference());
+            assertEquals(10, listJob.getPriority());
+            assertEquals(2, listJob.getRequestedWorkerSelectors().size());
+            assertEquals(Duration.ofSeconds(100), listJob.getRequestedWorkerSelectors().get(0).getExpiresAfter());
+        }
 
         List<RouterJobOffer> jobOffers = new ArrayList<RouterJobOffer>();
         long startTimeMillis = System.currentTimeMillis();
@@ -150,18 +205,39 @@ public class RouterJobLiveTests extends JobRouterTestBase {
         UnassignJobResult unassignJobResult = jobRouterClient.unassignJob(jobId, assignmentId);
 
         // Verify
-        assertEquals(1, unassignJobResult.getUnassignmentCount());
+        assertTrue(unassignJobResult.getUnassignmentCount() > 0);
 
+        if (this.getTestMode() != TestMode.PLAYBACK) {
+            Thread.sleep(5000);
+        }
+
+        RouterQueueStatistics queueStatistics = jobRouterClient.getQueueStatistics(queueId);
+
+        // Verify
+        assertEquals(queueId, queueStatistics.getQueueId());
+        assertEquals(1, queueStatistics.getLength());
+        assertTrue(queueStatistics.getLongestJobWaitTimeMinutes() > 0);
+        assertTrue(queueStatistics.getEstimatedWaitTime().get(10).getSeconds() > 0);
+
+        RouterQueueStatistics deserialized = jobRouterClient.getQueueStatisticsWithResponse(queueId, null)
+            .getValue().toObject(RouterQueueStatistics.class);
+        assertEquals(queueId, deserialized.getQueueId());
+        assertEquals(1, deserialized.getLength());
+        assertTrue(deserialized.getLongestJobWaitTimeMinutes() > 0);
+        assertTrue(deserialized.getEstimatedWaitTime().get(10).getSeconds() > 0);
+
+        // Cleanup
         RequestOptions requestOptions = new RequestOptions();
         CancelJobOptions cancelJobOptions = new CancelJobOptions()
             .setDispositionCode("dispositionCode")
             .setNote("note");
         requestOptions.setBody(BinaryData.fromObject(cancelJobOptions));
-
-        // Cleanup
         jobRouterClient.cancelJob(jobId, requestOptions);
         jobRouterClient.deleteJob(jobId);
         jobRouterClient.deleteWorker(workerId);
+        if (this.getTestMode() != TestMode.PLAYBACK) {
+            Thread.sleep(5000);
+        }
         routerAdminClient.deleteQueue(queueId);
         routerAdminClient.deleteDistributionPolicy(distributionPolicyId);
     }
@@ -189,11 +265,10 @@ public class RouterJobLiveTests extends JobRouterTestBase {
         assertEquals(RouterJobStatus.PENDING_SCHEDULE, job.getStatus());
 
         // Action
-        RouterJob job2 = jobRouterClient.updateJobWithResponse(jobId,
-            BinaryData.fromObject(new RouterJob().setMatchingMode(new QueueAndMatchMode())),
-            new RequestOptions()).getValue().toObject(RouterJob.class);
+        job.setMatchingMode(new QueueAndMatchMode());
+        RouterJob updatedJob = jobRouterClient.updateJob(jobId, BinaryData.fromObject(job), null).toObject(RouterJob.class);
 
-        assertEquals(RouterJobStatus.QUEUED, job2.getStatus());
+        assertEquals(RouterJobStatus.QUEUED, updatedJob.getStatus());
 
         // Cleanup
         jobRouterClient.cancelJob(jobId);